refactor: LocationMessage 의 패키지 수정
